Initial Commit for Freight
This commit sets up Freight as a way to learn how a package manager and
build tool like cargo works. This initial commit creates a bare bones
setup where we use a justfile to create our executable by hand through
invocations of rustc. This executable is then run to build itself again
with a different configuration to show that we can build the program
with itself, and then that executable is run to show that it was built
by itself successfully. Currently it just outputs that it did, but in
the future we would want to have it do a bit more to show it built fine
and can run.