Change RustcBuilder to use `impl Into` for fn args
This change makes functions for RustcBuilder use `impl Into` to make the
API a bit more convenient to use. This means we can pass in String or
PathBuf or &str or Path giving us a more flexible API with what we can
accept. This is much nicer to work with overall.