Add basic arg parsing for Freight
This does basic argument parsing for freight so that we can make freight
more useful by adding commands to actually run the program with and do
things with them. Following adding the command for printing out a help
message, we will add the build command so that we can well build freight
and other projects. Once we do we'll be able to remove the stage1 cfg
flags because we'll have a generic strategy for building things. The
only difference will be where it outputs it's build files which the
Justfile will take care of.