Make freight build itself w/o a bootstrap feature
This commit makes freight more useful by having it actually build itself
without the bootstrap feature. We do this by being a bit more generic
and any rust project with no dependencies or workspace setup with either
a `src/main.rs` or `src/lib.rs or some combination of the two will
compile now. We make quite a few assumptions to get this to work, but
that's fine as we will address this in future changes because we'll want
to start working on having a config file soon if not next.