Implement basic configuration w/Freight.toml file
This commit adds the beginning of a configuration file for Freight.
Eventually we'll make it more robust with parsing when we can add serde
and the like as a dependency but we're a long ways off from that. With
this we can offload configuration from being hardcoded inside of Freight
to a config file that users can use to tailor to their project. It also
lets us get rid of the janky check for ".git" in root_dir and instead
look for a manifest to work with, in this case Freight. Later on we'll
need better toml support so that we can build Cargo projects