Add FromStr impl for Edition and BoxError alias
This commit moves the parsing code we did for Edition in the Manifest
type to a FromStr impl so that we can instead use this in other places
as well if need be and have the parsing code for Manifest be more about
parsing out fields and not doing adhoc type conversions in that parsing
if we can move it into the type itself somehow. We also alias `Box<dyn Error>`
to `BoxError` so that we can easily use that for our Error type for now.