Create the Logger type to log data to the terminal
cargo has some fancy ways to print out things to the terminal and show
status. Eventually we'll get there as we add more things like
dependencies etc. However, we have to start somewhere with outputting
things in a way that's not just println!. This commit starts that
process by creating a Logger type that will print out what it's
compiling and when it's done!