Add testing apparatus to freight with freight test
We finally have tests that aren't just "Run the cli to make sure it
runs!" With this we can start adding more and more tests to make sure
that freight is well tested and catch things that might break a bit more
easily than we could before. We make quite a few changes here that are
important to cover:

- Expanded the derive traits for Edition and CrateType so that we can
  test them better
- Added support to parse from a string to CrateType, just like Edition
- Added tests to make sure that string parsing works as expected for
  them
- Moved the bin_compile and lib_compile function out to be fn() instead
  of a closure
- Added a test_compile function to compile libs and bins as tests
- Added a function to just build the tests
- Added a function to just run the tests
- Update CI to build and to run tests
- Update pre-commit hook to build and to run tests

This captures most of it. With these changes we can be more confident
that things work and can more thoroughly test our code as we go along.