Add freight doc command
This commit adds the ability to call rustdoc on freight projects. Prior
to this we called rustdoc in order to build and run tests and now we can
actually document our code which we have not been doing up to this point
since we could not open or read the docs.

A stub has been left for the open command to be done at another time.

This commit also denotes to possible future improvements for both
testing and documenting to handle cases when both main.rs and lib.rs
exist at the same time.