Add the freight init command
This command adds the ability to initialize a Freight project in the
current directory. Future work will let people specify directory to use.
We also include a test in tests/init.rs that fully makes sure that the
directory is initialized the way we expect and automatically does
cleanup for us.