Add the new command to freight and update help.txt
This adds the command `new` to freight which creates a brand new project
in a given directory. It is exactly like init, but you choose where to
initialize the project. We also update the help text to include both the
init command which had not been updated in the help text and the new
command.